machine/requirements.txt
flask==3.0.3
flask-cors==4.0.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.3.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Necessário para Pub/Sub (API e Worker)
google-cloud-pubsub==2.21.0
